capstone-ecommerce-microservices/customer-service/customer_prj/requirements.txt
Django>=5.1.7
djangorestframework>=3.14.0
snowflake-connector-python>=3.14.0
django-snowflake>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
django-cors-headers>=4.3.1
requests>=2.31.0
cryptography>=42.0.0